+++ api/stream-bytesex.h
@@ -65,7 +65,7 @@ int getBitFromByteStream(const void *byteStream, int bitIndex);        //< bitIn
 /// sets serval bits in a bigEndian bitfeld of max 16 bits (warning: ccsds: 0 = msb)
 void setBitField(void* buffer, size_t bitPos, uint8_t numOfBits, uint32_t val);
 /// gets severa bits from a bigEndia bitfled of max 16 bits (warning: ccsds: 0 = msb)
-uint32_t getBitField(const void* buffer, int bitPos, int numOfBits);
+uint32_t getBitField(const void* buffer, size_t bitPos, uint8_t numOfBits);
 
 /// gets the number of bits set
 uint32_t getSetBits(const uint32_t &value);
