+++ src/bare-metal/raspberrypi3/include/hw_tcp.h
@@ -1,5 +1,7 @@
 #pragma once
 
+#include <stdint.h>
+
 // DUMMY!
 
 namespace RODOS {
@@ -10,8 +12,8 @@ class HW_TCPServer {
 
     bool listen([[gnu::unused]] const long portNr) { return false; };
     bool acceptNewConnection() { return false; };
-    int  sendData([[gnu::unused]] void* buf, [[gnu::unused]] size_t len) { return 0; };
-    int  getData([[gnu::unused]] void* buf, [[gnu::unused]] size_t maxLen) { return 0; };
+    int  sendData([[gnu::unused]] void* buf, [[gnu::unused]] uint16_t len) { return 0; };
+    int  getData([[gnu::unused]] void* buf, [[gnu::unused]] uint16_t maxLen) { return 0; };
     int  getErrorCode() const { return -1; };
 };
 
@@ -20,8 +22,8 @@ class HW_TCPClient {
     ~HW_TCPClient();
 
     bool reopen([[gnu::unused]] const long portNr, [[gnu::unused]] const char* hostname = "localhost") { return false; };
-    int  sendData([[gnu::unused]] void* buf, [[gnu::unused]] size_t len) { return 0; };
-    int  getData([[gnu::unused]] void* buf, [[gnu::unused]] size_t maxLen) { return 0; };
+    int  sendData([[gnu::unused]] void* buf, [[gnu::unused]] uint16_t len) { return 0; };
+    int  getData([[gnu::unused]] void* buf, [[gnu::unused]] uint16_t maxLen) { return 0; };
     int  getErrorCode() const { return -1; };
 };
